requirements.txt
Flask
flask-cors
openpyxl
gunicorn
Pillow
requests
pyinstaller
reportlab